chore: remove circular imports
