+++ src/components/WindowsContainer.tsx
@@ -3,13 +3,11 @@ import React, { useRef } from "react";
 import { createPortal } from "react-dom";
 import { CSSTransition, TransitionGroup } from "react-transition-group";
 import { useWindowManager } from "../hooks";
-import { getFadeInAnimation } from "./getFadeInAnimation";
+import { defaultFadeAnimation } from "./getFadeInAnimation";
 import { ModalMask } from "./ModalMask";
 import { Window } from "./window/Window";
 import { ContentGetter } from "./window/WindowContent";
 import { WindowsViewport } from "./WindowsViewport";
-
-export const fadeInAnimation = getFadeInAnimation();
 
 interface WindowsContainerProps {
   container?: HTMLElement;
@@ -25,11 +23,11 @@ export function WindowsContainer({ container = document.body, contentGetter }: W
     <TransitionGroup component={WindowsViewport}>
       {flatMap(windows, (d, index) => [
         d.isModal && (
-          <CSSTransition nodeRef={modalMaskRef} key={`${d.id}/mask`} timeout={250} classNames={fadeInAnimation}>
+          <CSSTransition nodeRef={modalMaskRef} key={`${d.id}/mask`} timeout={250} classNames={defaultFadeAnimation}>
             <ModalMask ref={modalMaskRef} key={`${d.id}/mask`} zIndex={index} />
           </CSSTransition>
         ),
-        <CSSTransition nodeRef={windowRef} key={d.id} timeout={250} classNames={fadeInAnimation}>
+        <CSSTransition nodeRef={windowRef} key={d.id} timeout={250} classNames={defaultFadeAnimation}>
           <Window ref={windowRef} data={d} contentGetter={contentGetter} />
         </CSSTransition>,
       ]).filter(Boolean)}
